% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOptimality.R
\name{AEfficiency}
\alias{AEfficiency}
\title{Calculate the A-Efficiency for Designs without hard to change factors}
\usage{
AEfficiency(doeDesign)
}
\arguments{
\item{doeDesign}{An object of class \code{\link{doeDesign-class}} for which the A-Efficiency should be calculated.}
}
\value{
A numeric representing the A-efficiency of the design on a range from 0 to 100.
}
\description{
A-Efficiency expresses the quality of a design in terms of A-Optimality as a percentage. 100% A-Efficiency is the
best possible design (in terms of A-Optimality) for a given model and a given number of runs. A-Efficiency is calculated
as \eqn{100*(1/n*trace((X'X)^{-1})/p)}.
}
\details{
Currently there is no implementation for split-plot-designs (or any design that uses a hard to change or semi hard to
change factor).
}
\examples{
doe <- GenerateNewDoeDesign(
 factors = list(
  new("doeFactor", name="X1"),
  new("doeFactor", name="X2")
 ),
 number.runs = as.integer(20),
 design.model = ~X1*X2,
 optimality.function=AOptimality,
 optimality.criterion = "A-Optimality",
 random.doe = TRUE
)

# AEfficiency of the randon initial design
AEfficiency(doe)

# Aefficiency of the optimized design
doe.Aopt <- GenerateOptimalDesign(doe, 1, 10)
AEfficiency(doe.Aopt[[1]])

}
