% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateDesign.R
\name{AliasingHeatMap}
\alias{AliasingHeatMap}
\title{Draw a color map of correlations}
\usage{
AliasingHeatMap(doe, fml = NULL)
}
\arguments{
\item{doe}{the design to evaluate}

\item{fml}{a (optional) formula representing the design model of interest (including aliased terms)}
}
\description{
Draw a color map of correlations
}
\examples{
doe <- GenerateNewDoeDesign(
 factors = list(
  new("doeFactor", name="X1"),
  new("doeFactor", name="X2"),
  new("doeFactor", name="X3"),
  new("doeFactor", name="X4")
 ),
 number.runs = as.integer(8),
 design.model = ~X1+X2+X3+X4,
 optimality.function=DOptimality,
 optimality.criterion = "D-Optimality",
 random.doe = TRUE
)

# Generate a d-optimal design
doe.Dopt <- GenerateOptimalDesign(doe, 1, 10)
AliasingHeatMap(doe.Dopt[[1]], fml=~X1+X2+X3+X4)
a.matrix1 <- AliasingHeatMap(doe.Dopt[[1]], fml=~X1+X2+X3+X4+X1*X2+X1*X3+ X1*X4+X2*X3+X2*X4+X3*X4)
a.matrix1

# Generate a A-optimal design
doe@optimality.function<-AOptimality
doe.Aopt <- GenerateOptimalDesign(doe, 1, 10)
a.matrix2 <- AliasingHeatMap(doe.Aopt[[1]], fml=~X1+X2+X3+X4+X1*X2+X1*X3+ X1*X4+X2*X3+X2*X4+X3*X4)
a.matrix2

# Compare both designs (using the multiplot function from here: http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_\%28ggplot2\%29/)
multiplot <- function(..., plotlist=NULL, file, cols=1, layout=NULL) {
library(grid)
plots <- c(list(...), plotlist)
numPlots = length(plots)
if (is.null(layout)) {
  layout <- matrix(seq(1, cols * ceiling(numPlots/cols)),
                   ncol = cols, nrow = ceiling(numPlots/cols))
}
if (numPlots==1) {
  print(plots[[1]])

} else {
  # Set up the page
  grid.newpage()
  pushViewport(viewport(layout = grid.layout(nrow(layout), ncol(layout))))

  # Make each plot, in the correct location
  for (i in 1:numPlots) {
    matchidx <- as.data.frame(which(layout == i, arr.ind = TRUE))

    print(plots[[i]], vp = viewport(layout.pos.row = matchidx$row,
                                    layout.pos.col = matchidx$col))
  }
}
}
multiplot(a.matrix1, a.matrix2, cols=2)
}
