% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AOptimality.R
\name{AOptimality}
\alias{AOptimality}
\title{Calculate A Optimality for a design.}
\usage{
AOptimality(currentDesign)
}
\arguments{
\item{doeDesign}{the design to calculate the A-Optimality from-}
}
\description{
Calculate A Optimality for a design.
}
