% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doeDesign.R
\name{check_doeDesign}
\alias{check_doeDesign}
\title{Checks the validity of an object of class \code{\link{doeDesign-class}}.}
\usage{
check_doeDesign(object)
}
\description{
Checks the validity of an object of class \code{\link{doeDesign-class}}.
}
\keyword{internal}
