% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doeFactor.R
\name{check_doeFactor}
\alias{check_doeFactor}
\title{Check validity of doeFactor class}
\usage{
check_doeFactor(object)
}
\arguments{
\item{object}{An object of class doeFactor.}
}
\description{
Check validity of doeFactor class
}
\keyword{It}
\keyword{This}
\keyword{and}
\keyword{checking}
\keyword{declaration}
\keyword{doeFactor-objects.}
\keyword{does}
\keyword{during}
\keyword{error}
\keyword{for}
\keyword{function}
\keyword{if}
\keyword{internal}
\keyword{is}
\keyword{needed.}
\keyword{object}
\keyword{of}
\keyword{returns}
\keyword{some}
\keyword{the}
\keyword{used}
