% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doeFactor.R
\docType{class}
\name{doeFactor-class}
\alias{doeFactor-class}
\title{An S4 class representing a factor in a designed experiment.}
\description{
A factor is a parameter that is varied during the experimentation. This is not to confuse with R's factors that are categorical variables.
}
\section{Slots}{

\describe{
\item{\code{name}}{A character representing the name of the factor.}

\item{\code{levels}}{A vector containing the different levels of the factor. For continuous factors the minimum and maximum levels should be provided.}

\item{\code{number.levels}}{An integer representing the number of different levels a continuous factor can take. The number of levels and the 'levels' (minimum/maximum) define what levels a continuous factor can take. Design generation might be very slow for many factor levels.}

\item{\code{type}}{A character representing the modeling type of the factor. Accepted values are: continuous or categorical.}

\item{\code{changes}}{A character representing if a factor is 'easy', 'hard' or 'semi.hard' -to-change.}

\item{\code{semi.htc.group.size}}{An integer specifing how many different levels can be used in one block for a semi.hard-to-change-factor.}
}}

