% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DOptimality.R
\name{DOptimality}
\alias{DOptimality}
\title{Calculate the D-optimality of a design.}
\usage{
DOptimality(doeDesign)
}
\description{
This function calculates the D-Optimality of a design (\eqn{d_{opt} = |X'X|}, where \eqn{X} is the design matrix). D-optimality is
calculated for completely randomized designs or for split-plot-designs. For split plot designs D-Optimality is calculated based on Jones, Goos (2017):
\eqn{|M| = |X'V^{-1}X|} with V being the covariance matrix of the responses: \eqn{V=\sigma^2_{\epsilon} I_n + \sigma_{\gamma}^2ZZ'} and
\eqn{\sigma^2_{\epsilon}} represents the error variation and \eqn{\sigma^2_{\gamma}} represents the whole-plot-variation.
}
\examples{
splitplot <- GenerateNewDoeDesign(
 factors = list(
  new("doeFactor", name="X1", levels=c("A", "B", "C"), type="categorical", changes="hard"),
  new("doeFactor", name="X2", levels=c(0, 1))
 ),
 number.runs = as.integer(20),
 whole.plot.structure=data.frame(X1=rep(1:5, each=4)),
 design.model = ~X1*X2,
 optimality.function=DOptimality,
 optimality.criterion = "D-Optimality",
 random.doe = TRUE
)

splitplot.doe <- GenerateOptimalDesign(splitplot, random.start=1, max.iter=3)

DOptimality(splitplot.doe[[1]])
}
\references{
Bradley Jones & Peter Goos (2017) I-Optimal Versus D-Optimal Split-Plot Response Surface Designs, Journal of Quality Technology, 44:2, 85-101, DOI: 10.1080/00224065.2012.11917886
}
