% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateOptimalDesign.R
\name{GenerateOptimalDesign}
\alias{GenerateOptimalDesign}
\title{Algorithm to generate optimal designs.}
\usage{
GenerateOptimalDesign(doeSpec, random.start = 1, max.iter = 25,
  quick.shtc.search = TRUE, continue = FALSE, silent = FALSE)
}
\arguments{
\item{doeSpec}{An object of class \code{\link{doeDesign-class}} defining the attributes of the DOE.}

\item{random.start}{How many random starts shall be tried. For each random start one DOE will be generated. Depending on the complexity of the DOE (especially: number of factors and number of factor levels) the runtime of this function might increased by a lot when using many random starts.}

\item{max.iter}{The number of iterations that are used for each random start. One iteration is updating each cell in the DOE once. Depending on the complexity of the DOE (especially: number of factors and number of factor levels) the runtime of this function might increased by a lot when using many iterations.}

\item{quick.shtc.search}{(boolean, default is TRUE) Testing all possible combinations for SHTC factors in the updating of the design table can be extremly
time consuming. Quick shtc search follows a fast heuristic approach.}

\item{continue}{(boolean, default is FALSE) If TRUE the current design table is used as starting design instead of a
random design. If TRUE the argument random.start is ignored.}

\item{silent}{(boolean, default is FALSE) Should there be a status bar be printed in the console or not.}
}
\value{
A list with on optimized DOE for each random start. The list is ordered from best to worst in terms of optimality.
}
\description{
This function implements a variant of the candidate-free optimal algorithm described in Jones (2007).
}
\references{
Jones, B., & Goos, P. (2007). A candidate-set-free algorithm for generating D-optimal split-plot designs. Journal of the Royal Statistical Society. Series C, Applied Statistics, 56(3), 347–364. http://doi.org/10.1111/j.1467-9876.2007.00581.x
}
