% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doeDesign.R
\name{getDesignMatrix}
\alias{getDesignMatrix}
\title{This function returns the design matrix of a doeDesign.}
\usage{
getDesignMatrix(object, includeWholePlots = TRUE)
}
\arguments{
\item{includeWholePlots}{(logical, default=TRUE) if TRUE the data frame contains the whole plot structure as part of the data frame.}
}
\description{
This function returns the design matrix of a doeDesign.
}
