% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doeDesign.R
\name{getEtcFactors}
\alias{getEtcFactors}
\title{Function that returns a list with all easy-to-change factors of a DOE.}
\usage{
getEtcFactors(doeD)
}
\arguments{
\item{doeD}{An object of class \code{\link{doeDesign-class}}.}
}
\value{
A list of \code{\link{doeFactor-class}} in the DOE that are easy to change.
}
\description{
Mostly for internal use.
}
