% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doeDesign.R
\name{getFactorByName}
\alias{getFactorByName}
\title{Function that returns a factor in a doe identifying it via its name.}
\usage{
getFactorByName(doe, factor.name)
}
\arguments{
\item{doe}{An object of class \code{\link{doeDesign-class}}.}

\item{factor.name}{A character representing the name of the factor of interest.}
}
\description{
Function that returns a factor in a doe identifying it via its name.
}
\keyword{internal}
