% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doeDesign.R
\name{getFactorNames}
\alias{getFactorNames}
\title{Function that returns a list containing the names of all factors of a DOE.}
\usage{
getFactorNames(object)
}
\arguments{
\item{object}{An object of class \code{\link{doeDesign-class}}.}
}
\description{
Function that returns a list containing the names of all factors of a DOE.
}
