% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateDesign.R
\name{getHierarchyLevel}
\alias{getHierarchyLevel}
\title{Get the hirarchie level of a model effect by its name}
\usage{
getHierarchyLevel(effectname)
}
\arguments{
\item{effectname}{}
}
\description{
Get the hirarchie level of a model effect by its name
}
