% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doeDesign.R
\name{getHtcFactors}
\alias{getHtcFactors}
\title{Function that returns a list with all hard-to-change factors of a DOE.}
\usage{
getHtcFactors(doeD)
}
\arguments{
\item{doeD}{An object of class \code{\link{doeDesign-class}}.}
}
\value{
A list of \code{\link{doeFactor-class}} in the DOE that are hard to change.
}
\description{
Mostly for internal use.
}
