% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doeFactor.R
\name{getPossibleLevels}
\alias{getPossibleLevels}
\title{Function that returns a list with all possible levels of a factor.}
\usage{
getPossibleLevels(object)
}
\arguments{
\item{object}{An object of class doeFactor.}
}
\description{
For categoric factors this is just the list of all levels. For continuous factors
the list is determined by the minimum and maximum levels provided in the slot levels
and the number of levels-slot of the factor.
}
\examples{
newFactor <- new("doeFactor",
                     name="Temperature",
                     levels=c(180, 210),
                     number.levels=as.integer(5),
                     type="continuous",
                     changes="easy")
getPossibleLevels(newFactor)
}
\keyword{internal}
