% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doeDesign.R
\name{getWholePlotColumn}
\alias{getWholePlotColumn}
\title{A function that returns the correpsonding whole plot structure column for a (semi-)hard to change factor.}
\usage{
getWholePlotColumn(doe, factor.name)
}
\description{
A function that returns the correpsonding whole plot structure column for a (semi-)hard to change factor.
}
\keyword{internal}
