% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveDOEtoJMP.R
\name{hlpr.JMP.model.script}
\alias{hlpr.JMP.model.script}
\title{Function that creates the design model script part for JMP.}
\usage{
hlpr.JMP.model.script(doe)
}
\description{
Function that creates the design model script part for JMP.
}
\keyword{internal}
