% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveDOEtoJMP.R
\name{hlpr.JMP.new.column.random}
\alias{hlpr.JMP.new.column.random}
\title{Function that creates the individual columns of the
whole plot effects in the data table based on the r-doe.}
\usage{
hlpr.JMP.new.column.random(doe)
}
\description{
Function that creates the individual columns of the
whole plot effects in the data table based on the r-doe.
}
\keyword{internal}
