% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveDOEtoJMP.R
\name{hlpr.JMP.new.column}
\alias{hlpr.JMP.new.column}
\title{Function that creates the JMP code to create the
columns representing the factors of the DOE.}
\usage{
hlpr.JMP.new.column(doe, index)
}
\description{
Function that creates the JMP code to create the
columns representing the factors of the DOE.
}
\keyword{internal}
