% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveDOEtoJMP.R
\name{hlpr.JMP.single.doe}
\alias{hlpr.JMP.single.doe}
\title{Collects all information of one doe and generates a JMP script
that creates the corresponding data set in JMP.}
\usage{
hlpr.JMP.single.doe(doe)
}
\description{
Collects all information of one doe and generates a JMP script
that creates the corresponding data set in JMP.
}
\keyword{internal}
