% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOptimality.R
\name{IOptimality}
\alias{IOptimality}
\title{Calculates the i-optimality of a design.}
\usage{
IOptimality(currentDesign, momentsMatrix, blockedVar)
}
\arguments{
\item{doe}{...}
}
\description{
I-optimality is the average prediction variance of a design. It is calculated using
the vdg-package.
}
