% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doeDesign.R
\name{reorderFactors}
\alias{reorderFactors}
\title{Function to reorder the factors in a DOE from hard to change to easy to change.}
\usage{
reorderFactors(doe)
}
\arguments{
\item{doe}{An object of class \code{\link{doeDesign-class}}.}
}
\value{
The same object that was passed as argument just with sorted factors.
}
\description{
This function is supposed to be used internally. Mostly the \code{\link{GenerateOptimalDesign}}-functions
expects that the factors are sorted from hard to change to easy to change. This happens automatically
when a design is created using \code{\link{GenerateNewDoeDesign}}.
}
