% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SaveDOEtoJMP.R
\name{SaveDOEtoJMP}
\alias{SaveDOEtoJMP}
\title{Function to export a DOE stored in an object of class \code{\link{doeDesign-class}} to JMP.}
\usage{
SaveDOEtoJMP(doe, filepath)
}
\arguments{
\item{doe}{A list of DOEs that are being exported.}

\item{filepath}{A character representing the path to store the JMP script.}
}
\description{
This function generates a JMP script that will produce a data table with the most relevant information
about the DOE. That includes the design table and model and the factor attributes as well.
}
