% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/doeFactor.R
\docType{methods}
\name{show,doeFactor-method}
\alias{show,doeFactor-method}
\title{Default print function for an object of class doeFactor.}
\usage{
\S4method{show}{doeFactor}(object)
}
\arguments{
\item{object}{An object of class doeFactor. A basic overview of this object will be printed to the console.}
}
\description{
Default print function for an object of class doeFactor.
}
\examples{
newFactor <- new("doeFactor",
                     name="Temperature",
                     levels=c(180, 210),
                     number.levels=as.integer(5),
                     type="continuous",
                     changes="easy")
newFactor
}
