% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateDesign.R
\name{sort.modeleffects}
\alias{sort.modeleffects}
\title{Helper function to order the model effects in the aliasing heatmap}
\usage{
\method{sort}{modeleffects}(effectnames)
}
\arguments{
\item{effectnames}{a vector of strings representing the effects
visualized in the aliasing heatmap.}
}
\description{
Helper function to order the model effects in the aliasing heatmap
}
